test: split adapter tests by source module
