+++ examples/basic/package.json
@@ -8,7 +8,7 @@
   "scripts": {
     "dev": "react-router dev",
     "build": "react-router build",
-    "start": "NODE_ENV=production node --experimental-strip-types ./server.ts",
+    "start": "NODE_ENV=production node ./server.ts",
     "typecheck": "react-router typegen && tsc"
   },
   "dependencies": {
